test: :zap: Amending the test timing for perf
Amending the test times to be on push to main via pr merge and on every pr in dev. So code commited will be checked in dev and the once again when being merged in main without doubling testing going on. Before when a PR Merge was being made, it would test the push to main and the pr resulting in two tests for the same code.